// components/ServiceMap/types.ts
export enum EdgeType {
  bezier = 'bezier',
  simplebezier = 'simplebezier',
  smoothstep = 'smoothstep',
  step = 'step',
  straight = 'straight',
}

export enum Orientation {
  horizontal = 'horizontal',
  vertical = 'vertical',
}

export enum ServiceMapType {
  cluster = 'Cluster',
  flow = 'Flow',
}